% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnorm.R
\docType{package}
\name{ggnorm}
\alias{ggnorm}
\alias{ggnorm-package}
\title{ggnorm: A package for drawing normal quantile plots with lines on ggplot}
\description{
ggnorm: A package for drawing normal quantile plots with lines on ggplot
}

